app terminada

+++ app/src/main/java/arquitectura/proyecto/android/appsgpl/POJOS/Proyecto.java
@@ -25,13 +25,13 @@ public class Proyecto implements Parcelable, Serializable {
     private String dateEndFake;
     private float monto;
 
-    public Proyecto(int idProyecto,String dateEndFake){
+    public Proyecto(int idProyecto, String dateEndFake){
         this.idProyecto=idProyecto;
         this.dateEndFake=dateEndFake;
     }
 
-    public Proyecto(int idEmpresa,int idCategoriaP,String nombreProyecto,String codigoProyecto,String descripcionProyecto
-    ,String dateStart,String dateEnd,int monto){
+    public Proyecto(int idEmpresa, int idCategoriaP, String nombreProyecto, String codigoProyecto, String descripcionProyecto
+            , String dateStart, String dateEnd, float monto){
         this.idEmpresa=idEmpresa;
         this.idCategoriaP=idCategoriaP;
         this.nombreProyecto=nombreProyecto;
